\name{Y}
\alias{Y}
\docType{data}
\title{
Observation matrix Y
}
\description{
An example of observation matrix
}
\usage{data("Y")}
\format{
  The format is:
 num [1:50] 9 2 11 14 18 17 1 0 1 0 ...
}
\references{
M. Gomtsyan "Variable selection in a specific regression time series of counts.", arXiv:2307.00929
}
\examples{
data(Y)
}
